+++ exec/plugin-commands-script-runners/src/dlx.ts
@@ -12,8 +12,8 @@ import { add } from '@pnpm/plugin-commands-installation'
 import { readPackageJsonFromDir } from '@pnpm/read-package-json'
 import { getBinsFromPackageManifest } from '@pnpm/package-bins'
 import { pickRegistryForPackage } from '@pnpm/pick-registry-for-package'
+import { type PnpmSettings } from '@pnpm/types'
 import execa from 'execa'
-import omit from 'ramda/src/omit'
 import pick from 'ramda/src/pick'
 import renderHelp from 'render-help'
 import symlinkDir from 'symlink-dir'
@@ -75,7 +75,7 @@ export type DlxCommandOptions = {
   package?: string[]
   shellMode?: boolean
   allowBuild?: string[]
-} & Pick<Config, 'extraBinPaths' | 'registries' | 'reporter' | 'userAgent' | 'cacheDir' | 'dlxCacheMaxAge' | 'useNodeVersion' | 'symlink'> & add.AddCommandOptions
+} & Pick<Config, 'extraBinPaths' | 'registries' | 'reporter' | 'userAgent' | 'cacheDir' | 'dlxCacheMaxAge' | 'useNodeVersion' | 'symlink'> & add.AddCommandOptions & PnpmSettings
 
 export async function handler (
   opts: DlxCommandOptions,
@@ -108,22 +108,17 @@ export async function handler (
   if (!cacheExists) {
     fs.mkdirSync(cachedDir, { recursive: true })
     await add.handler({
-      // Ideally the config reader should ignore these settings when the dlx command is executed.
-      // This is a temporary solution until "@pnpm/config" is refactored.
-      ...omit(['workspaceDir', 'rootProjectManifest', 'symlink'], opts),
+      ...opts,
       bin: path.join(cachedDir, 'node_modules/.bin'),
       dir: cachedDir,
       lockfileDir: cachedDir,
-      rootProjectManifestDir: cachedDir, // This property won't be used as rootProjectManifest will be undefined
-      rootProjectManifest: {
-        pnpm: {
-          onlyBuiltDependencies: [...resolvedPkgAliases, ...(opts.allowBuild ?? [])],
-        },
-      },
+      onlyBuiltDependencies: [...resolvedPkgAliases, ...(opts.allowBuild ?? [])],
       saveProd: true, // dlx will be looking for the package in the "dependencies" field!
       saveDev: false,
       saveOptional: false,
       savePeer: false,
+      symlink: true,
+      workspaceDir: undefined,
     }, resolvedPkgs)
     try {
       await symlinkDir(cachedDir, cacheLink, { overwrite: true })