+++ cli/cli-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/cli-utils
 
+## 1000.0.12
+
+### Patch Changes
+
+- Updated dependencies [1c2eb8c]
+  - @pnpm/config@1002.4.1
+  - @pnpm/default-reporter@1001.3.3
+
 ## 1000.0.11
 
 ### Patch Changes