+++ cli/default-reporter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/default-reporter
 
+## 1001.3.3
+
+### Patch Changes
+
+- Updated dependencies [1c2eb8c]
+  - @pnpm/config@1002.4.1
+
 ## 1001.3.2
 
 ### Patch Changes