+++ store/plugin-commands-store/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/plugin-commands-store
 
+## 1000.0.13
+
+### Patch Changes
+
+- Updated dependencies [1c2eb8c]
+  - @pnpm/config@1002.4.1
+  - @pnpm/cli-utils@1000.0.12
+  - @pnpm/store-connection-manager@1000.0.11
+
 ## 1000.0.12
 
 ### Patch Changes