+++ workspace/manifest-writer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/workspace.manifest-writer
 
+## 1000.0.1
+
+### Patch Changes
+
+- 23754c7: Fix the update of `pnpm-workspace.yaml` by the `pnpm approve-builds` command [#9168](https://github.com/pnpm/pnpm/issues/9168).
+
 ## 1000.0.0
 
 ### Major Changes